requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
aiohttp>=3.9.0

# ArXiv
arxiv>=2.0.0

# Hugging Face
huggingface-hub>=0.20.0

# Social Media
tweepy>=4.14.0           # Twitter/X API
praw>=7.7.0              # Reddit API
PyGithub>=2.1.0          # GitHub API

# Data Processing
beautifulsoup4>=4.12.0
lxml>=5.0.0
feedparser>=6.0.0

# Async Support
asyncio-throttle>=1.0.0

# Utilities
rich>=13.0.0             # 美化输出
tenacity>=8.2.0          # 重试机制
python-dateutil>=2.8.0
numpy>=1.24.0

# Phase 2: Processing & Storage
qdrant-client>=1.7.0           # Qdrant 向量数据库 (支持元数据过滤、持久化、并发)

# Embedding APIs (免费)
# 使用 httpx 调用 SiliconFlow BGE-M3 API (无需额外依赖)
# 注册获取免费 API Key: https://siliconflow.cn/

# Optional: Local Embeddings
# sentence-transformers>=2.2.0   # 本地 SentenceTransformer
# transformers>=4.30.0           # HuggingFace Embeddings
# torch>=2.0.0                   # PyTorch (for transformers)

# Phase 3: Intelligence Layer (LLM + LangGraph)
langgraph>=0.2.0               # LangGraph 编排框架
openai>=1.0.0                  # OpenAI SDK (也用于 DeepSeek)
anthropic>=0.40.0              # Anthropic Claude SDK
google-generativeai>=0.8.0     # Google Gemini SDK

# Optional: LangChain Extensions
# langchain>=0.3.0
# langchain-openai>=0.2.0
# langchain-anthropic>=0.2.0

# Dev / Tests
pytest>=7.0.0
pytest-asyncio>=0.23.0

